Move the `use_duration_signed_ser`/`*_de` macros utils
This allows to use the same macro definition for all three time
libraries, and makes it easier to maintain the code.
